pyproject.toml
[project]
name = "ddbcoding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dolphindb>=3.0.4.2",
    "lancedb>=0.27.1",
    "llama-index>=0.14.13",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-retrievers-bm25>=0.6.5",
    "llama-index-vector-stores-lancedb>=0.4.4",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "uvicorn>=0.40.0",
]
